Time: 36 ms (94.23%) | Memory: 18.3 MB (72.25%) - LeetSync

+++ 637-average-of-levels-in-binary-tree/average-of-levels-in-binary-tree.py
@@ -8,7 +8,6 @@ class Solution:
     def averageOfLevels(self, root: Optional[TreeNode]) -> List[float]:
         if not root:
             return
-        
 
         queue = deque([root])
         res = []
@@ -23,6 +22,6 @@ class Solution:
                     queue.append(node.left)
                 if node.right:
                     queue.append(node.right)
-            avg = float(currsum/total_nodes)
+            avg = currsum / total_nodes
             res.append(avg)
         return res
